fixup! stream: implement toStreamIterWriter/fromStreamIterWriter
