+++ lib/internal/streams/iter/from.js
@@ -37,7 +37,7 @@ const {
 } = require('internal/util/types');
 
 const {
-  kTrustedSource,
+  kValidatedSource,
   toStreamable,
   toAsyncStreamable,
 } = require('internal/streams/iter/types');
@@ -484,8 +484,8 @@ function from(input) {
     throw new ERR_INVALID_ARG_TYPE('input', 'a non-null value', input);
   }
 
-  // Fast path: trusted source already yields valid Uint8Array[] batches
-  if (input[kTrustedSource]) {
+  // Fast path: validated source already yields valid Uint8Array[] batches
+  if (input[kValidatedSource]) {
     return input;
   }
 
@@ -538,17 +538,17 @@ function from(input) {
   // iteration protocols)
   if (typeof input[toAsyncStreamable] === 'function') {
     const result = input[toAsyncStreamable]();
-    // Synchronous trusted source (e.g. Readable batched iterator)
-    if (result?.[kTrustedSource]) {
+    // Synchronous validated source (e.g. Readable batched iterator)
+    if (result?.[kValidatedSource]) {
       return result;
     }
     return {
       __proto__: null,
       async *[SymbolAsyncIterator]() {
-        // The result may be a Promise. Check trusted on both the Promise
+        // The result may be a Promise. Check validated on both the Promise
         // itself (if tagged) and the resolved value.
         const resolved = await result;
-        if (resolved?.[kTrustedSource]) {
+        if (resolved?.[kValidatedSource]) {
           yield* resolved[SymbolAsyncIterator]();
           return;
         }
