+++ lib/internal/streams/iter/pull.js
@@ -51,7 +51,7 @@ const {
 } = require('internal/streams/iter/utils');
 
 const {
-  kTrustedTransform,
+  kValidatedTransform,
 } = require('internal/streams/iter/types');
 
 // =============================================================================
@@ -564,12 +564,12 @@ async function* applyStatefulAsyncTransform(source, transform, options) {
 }
 
 /**
- * Fast path for trusted stateful transforms (e.g. compression).
+ * Fast path for validated stateful transforms (e.g. compression).
  * Skips withFlushAsync (transform handles done internally) and
  * skips isUint8ArrayBatch validation (transform guarantees valid output).
  * @yields {Uint8Array[]}
  */
-async function* applyTrustedStatefulAsyncTransform(source, transform, options) {
+async function* applyValidatedStatefulAsyncTransform(source, transform, options) {
   const output = transform(source, options);
   for await (const batch of output) {
     if (batch.length > 0) {
@@ -639,8 +639,8 @@ async function* createAsyncPipeline(source, transforms, signal) {
         statelessRun = [];
       }
       const opts = { __proto__: null, signal: transformSignal };
-      if (transform[kTrustedTransform]) {
-        current = applyTrustedStatefulAsyncTransform(
+      if (transform[kValidatedTransform]) {
+        current = applyValidatedStatefulAsyncTransform(
           current, transform.transform, opts);
       } else {
         current = applyStatefulAsyncTransform(
