+++ lib/internal/streams/iter/transform.js
@@ -37,7 +37,7 @@ const {
 } = require('internal/errors');
 const { lazyDOMException } = require('internal/util');
 const { isArrayBufferView, isAnyArrayBuffer } = require('internal/util/types');
-const { kTrustedTransform } = require('internal/streams/iter/types');
+const { kValidatedTransform } = require('internal/streams/iter/types');
 const {
   checkRangesOrGetDefault,
   validateFiniteNumber,
@@ -306,7 +306,7 @@ function createZstdHandle(mode, options, processCallback, onError) {
 function makeZlibTransform(createHandleFn, processFlag, finishFlag) {
   return {
     __proto__: null,
-    [kTrustedTransform]: true,
+    [kValidatedTransform]: true,
     transform: async function*(source, options) {
       const { signal } = options;
