+++ lib/internal/streams/iter/types.js
@@ -45,30 +45,30 @@ const shareSyncProtocol = SymbolFor('Stream.shareSyncProtocol');
 const drainableProtocol = SymbolFor('Stream.drainableProtocol');
 
 /**
- * Internal sentinel for trusted stateful transforms. A transform object
- * with [kTrustedTransform] = true signals that:
+ * Internal sentinel for validated stateful transforms. A transform object
+ * with [kValidatedTransform] = true signals that:
  *   1. It handles source exhaustion (done) internally - no withFlushAsync
  *      wrapper needed.
  *   2. It always yields valid Uint8Array[] batches - no isUint8ArrayBatch
  *      validation needed on each yield.
  * This is NOT a public protocol symbol - it uses Symbol() not Symbol.for().
  */
-const kTrustedTransform = Symbol('kTrustedTransform');
+const kValidatedTransform = Symbol('kValidatedTransform');
 
 /**
- * Internal sentinel for trusted sources. An async iterable with
- * [kTrustedSource] = true signals that it already yields valid
+ * Internal sentinel for validated sources. An async iterable with
+ * [kValidatedSource] = true signals that it already yields valid
  * Uint8Array[] batches - no normalizeAsyncSource wrapper needed.
  * from() will return such sources directly, skipping all normalization.
  * This is NOT a public protocol symbol - it uses Symbol() not Symbol.for().
  */
-const kTrustedSource = Symbol('kTrustedSource');
+const kValidatedSource = Symbol('kValidatedSource');
 
 module.exports = {
   broadcastProtocol,
   drainableProtocol,
-  kTrustedSource,
-  kTrustedTransform,
+  kValidatedSource,
+  kValidatedTransform,
   shareProtocol,
   shareSyncProtocol,
   toAsyncStreamable,
