+++ lib/internal/streams/readable.js
@@ -30,15 +30,12 @@ const {
   ObjectKeys,
   ObjectSetPrototypeOf,
   Promise,
-  PromisePrototypeThen,
-  PromiseWithResolvers,
   ReflectApply,
   SafeSet,
   Symbol,
   SymbolAsyncDispose,
   SymbolAsyncIterator,
   SymbolFor,
-  SymbolIterator,
   SymbolSpecies,
   TypedArrayPrototypeSet,
 } = primordials;
@@ -96,7 +93,6 @@ const {
 } = require('internal/errors');
 const {
   validateAbortSignal,
-  validateInteger,
   validateObject,
 } = require('internal/validators');
 
@@ -1805,276 +1801,41 @@ Readable.wrap = function(src, options) {
   }).wrap(src);
 };
 
-// Efficient interop with the stream/iter API via toAsyncStreamable protocol.
-// Provides a batched async iterator that drains the internal buffer into
-// Uint8Array[] batches, avoiding the per-chunk Promise overhead of the
-// standard Symbol.asyncIterator path.
+// Interop with the stream/iter API via the toAsyncStreamable protocol.
+//
+// The batched iterator logic lives in classic.js (shared with the
+// fromReadable() utility for duck-typed streams). This prototype method
+// calls createBatchedAsyncIterator directly -- it must NOT call
+// fromReadable() since fromReadable() checks for toAsyncStreamable,
+// which would create infinite recursion.
 //
 // The flag cannot be checked at module load time (readable.js loads during
 // bootstrap before options are available). Instead, toAsyncStreamable is
-// always defined but lazily initializes on first call - throwing if the
-// flag is not set, or installing the real implementation if it is.
+// always defined but lazily initializes on first call -- throwing if the
+// flag is not set.
 {
-  const kNullPrototype = { __proto__: null };
   const toAsyncStreamable = SymbolFor('Stream.toAsyncStreamable');
-  let kTrustedSource;
-  let normalizeAsyncValue;
-  let isU8;
-
-  // Maximum chunks to drain into a single batch. Bounds peak memory when
-  // _read() synchronously pushes many chunks into the buffer.
-  const MAX_DRAIN_BATCH = 128;
-
-  function lazyInit() {
-    if (kTrustedSource !== undefined) return;
-    if (!getOptionValue('--experimental-stream-iter')) {
-      throw new ERR_STREAM_ITER_MISSING_FLAG();
-    }
-    ({ kTrustedSource } = require('internal/streams/iter/types'));
-    ({ normalizeAsyncValue } = require('internal/streams/iter/from'));
-    ({ isUint8Array: isU8 } = require('internal/util/types'));
-  }
-
-  // Normalize a batch of raw chunks from an object-mode or encoded
-  // Readable into Uint8Array values. Returns the normalized batch,
-  // or null if normalization produced no output.
-  async function normalizeBatch(raw) {
-    const batch = [];
-    for (let i = 0; i < raw.length; i++) {
-      const value = raw[i];
-      if (isU8(value)) {
-        batch.push(value);
-      } else {
-        // normalizeAsyncValue may await for async protocols (e.g.
-        // toAsyncStreamable on yielded objects). Stream events during
-        // the suspension are queued, not lost - errors will surface
-        // on the next loop iteration after this yield completes.
-        for await (const normalized of normalizeAsyncValue(value)) {
-          batch.push(normalized);
-        }
-      }
-    }
-    return batch.length > 0 ? batch : null;
-  }
-
-  // Batched async iterator for Readable streams. Same mechanism as
-  // createAsyncIterator (same event setup, same stream.read() to
-  // trigger _read(), same teardown) but drains all currently buffered
-  // chunks into a single Uint8Array[] batch per yield, amortizing the
-  // Promise/microtask cost across multiple chunks.
-  //
-  // When normalize is provided (object-mode / encoded streams), each
-  // drained batch is passed through it to convert chunks to Uint8Array.
-  // When normalize is null (byte-mode), chunks are already Buffers
-  // (Uint8Array subclass) and are yielded directly.
-  async function* createBatchedAsyncIterator(stream, normalize) {
-    let callback = nop;
-
-    function next(resolve) {
-      if (this === stream) {
-        callback();
-        callback = nop;
-      } else {
-        callback = resolve;
-      }
-    }
-
-    stream.on('readable', next);
-
-    let error;
-    const cleanup = eos(stream, { writable: false }, (err) => {
-      error = err ? aggregateTwoErrors(error, err) : null;
-      callback();
-      callback = nop;
-    });
-
-    try {
-      while (true) {
-        const chunk = stream.destroyed ? null : stream.read();
-        if (chunk !== null) {
-          // Drain any additional already-buffered chunks into the same
-          // batch. The first read() may trigger _read() which
-          // synchronously pushes more data into the buffer. We drain
-          // that buffered data without issuing unbounded _read() calls -
-          // once state.length hits 0 or MAX_DRAIN_BATCH is reached, we
-          // stop and yield what we have.
-          const batch = [chunk];
-          while (batch.length < MAX_DRAIN_BATCH &&
-                 stream._readableState.length > 0) {
-            const c = stream.read();
-            if (c === null) break;
-            batch.push(c);
-          }
-          if (normalize !== null) {
-            const result = await normalize(batch);
-            if (result !== null) {
-              yield result;
-            }
-          } else {
-            yield batch;
-          }
-        } else if (error) {
-          throw error;
-        } else if (error === null) {
-          return;
-        } else {
-          await new Promise(next);
-        }
-      }
-    } catch (err) {
-      error = aggregateTwoErrors(error, err);
-      throw error;
-    } finally {
-      if (error === undefined || stream._readableState.autoDestroy) {
-        destroyImpl.destroyer(stream, null);
-      } else {
-        stream.off('readable', next);
-        cleanup();
-      }
-    }
-  }
+  let createBatchedAsyncIterator;
+  let normalizeBatch;
+  let kValidatedSource;
 
   Readable.prototype[toAsyncStreamable] = function() {
-    lazyInit();
+    if (createBatchedAsyncIterator === undefined) {
+      if (!getOptionValue('--experimental-stream-iter')) {
+        throw new ERR_STREAM_ITER_MISSING_FLAG();
+      }
+      ({
+        createBatchedAsyncIterator,
+        normalizeBatch,
+      } = require('internal/streams/iter/classic'));
+      ({ kValidatedSource } = require('internal/streams/iter/types'));
+    }
     const state = this._readableState;
     const normalize = (state.objectMode || state.encoding) ?
-      normalizeBatch :
-      null;
+      normalizeBatch : null;
     const iter = createBatchedAsyncIterator(this, normalize);
-    iter[kTrustedSource] = true;
+    iter[kValidatedSource] = true;
     iter.stream = this;
     return iter;
-  };
-
-  // Create a byte-mode Readable from an AsyncIterable<Uint8Array[]>.
-  // The source must yield Uint8Array[] batches (the stream/iter native
-  // format). Each Uint8Array in a batch is pushed as a separate chunk.
-  Readable.fromStreamIter = function fromStreamIter(source, options = kNullPrototype) {
-    lazyInit();
-    if (typeof source?.[SymbolAsyncIterator] !== 'function') {
-      throw new ERR_INVALID_ARG_TYPE('source', 'AsyncIterable', source);
-    }
-
-    validateObject(options, 'options');
-    const {
-      highWaterMark = 64 * 1024,
-      signal,
-    } = options;
-    validateInteger(highWaterMark, 'options.highWaterMark', 0);
-    if (signal !== undefined) {
-      validateAbortSignal(signal, 'options.signal');
-    }
-
-    const iterator = source[SymbolAsyncIterator]();
-    let backpressure;
-    let pumping = false;
-    let done = false;
-
-    const readable = new Readable({
-      __proto__: null,
-      highWaterMark,
-      read() {
-        if (backpressure) {
-          const { resolve } = backpressure;
-          backpressure = null;
-          resolve();
-        } else if (!pumping && !done) {
-          pumping = true;
-          pump();
-        }
-      },
-      destroy(err, cb) {
-        done = true;
-        // Wake up the pump if it's waiting on backpressure so it
-        // can see done === true and exit cleanly.
-        if (backpressure) {
-          backpressure.resolve();
-          backpressure = null;
-        }
-        if (typeof iterator.return === 'function') {
-          PromisePrototypeThen(iterator.return(),
-                               () => cb(err), (e) => cb(e || err));
-        } else {
-          cb(err);
-        }
-      },
-    });
-
-    if (signal) {
-      addAbortSignalNoValidate(signal, readable);
-    }
-
-    async function pump() {
-      try {
-        while (!done) {
-          const { value: batch, done: iterDone } = await iterator.next();
-          if (iterDone) {
-            done = true;
-            readable.push(null);
-            return;
-          }
-          // Individual chunks are not validated as Uint8Array here.
-          // The caller is responsible for providing a well-formed
-          // AsyncIterable<Uint8Array[]> source. If a non-Buffer chunk
-          // is pushed, the byte-mode Readable will throw internally.
-          for (let i = 0; i < batch.length; i++) {
-            if (!readable.push(batch[i])) {
-              // Backpressure: wait for next _read() call
-              backpressure = PromiseWithResolvers();
-              await backpressure.promise;
-              if (done) return;  // Destroyed while waiting
-            }
-          }
-        }
-      } catch (err) {
-        done = true;
-        readable.destroy(err);
-      }
-    }
-
-    return readable;
-  };
-
-  // Create a byte-mode Readable from an Iterable<Uint8Array[]>.
-  // Fully synchronous - _read() pulls from the iterator directly.
-  // The source must yield Uint8Array[] batches.
-  Readable.fromStreamIterSync = function fromStreamIterSync(source, options = kNullPrototype) {
-    lazyInit();
-    if (typeof source?.[SymbolIterator] !== 'function') {
-      throw new ERR_INVALID_ARG_TYPE('source', 'Iterable', source);
-    }
-
-    validateObject(options, 'options');
-    const {
-      highWaterMark = 64 * 1024,
-    } = options;
-    validateInteger(highWaterMark, 'options.highWaterMark', 0);
-
-    const iterator = source[SymbolIterator]();
-
-    return new Readable({
-      __proto__: null,
-      highWaterMark,
-      read() {
-        for (;;) {
-          const { value: batch, done } = iterator.next();
-          if (done) {
-            this.push(null);
-            return;
-          }
-          // Individual chunks are not validated as Uint8Array here.
-          // The caller is responsible for providing a well-formed
-          // Iterable<Uint8Array[]> source. If a non-Buffer chunk
-          // is pushed, the byte-mode Readable will throw internally.
-          for (let i = 0; i < batch.length; i++) {
-            if (!this.push(batch[i])) return;
-          }
-        }
-      },
-      destroy(err, cb) {
-        if (typeof iterator.return === 'function') iterator.return();
-        cb(err);
-      },
-    });
   };
 }
