+++ lib/internal/streams/writable.js
@@ -26,26 +26,17 @@
 'use strict';
 
 const {
-  ArrayIsArray,
   ArrayPrototypeSlice,
   Error,
   FunctionPrototypeSymbolHasInstance,
-  MathMax,
-  NumberMAX_SAFE_INTEGER,
   ObjectDefineProperties,
   ObjectDefineProperty,
   ObjectSetPrototypeOf,
   Promise,
-  PromisePrototypeThen,
-  PromiseReject,
-  PromiseResolve,
-  PromiseWithResolvers,
   StringPrototypeToLowerCase,
   Symbol,
   SymbolAsyncDispose,
-  SymbolDispose,
   SymbolHasInstance,
-  TypedArrayPrototypeGetByteLength,
 } = primordials;
 
 module.exports = Writable;
@@ -57,16 +48,8 @@ const { Buffer } = require('buffer');
 const destroyImpl = require('internal/streams/destroy');
 const { eos } = require('internal/streams/end-of-stream');
 const {
-  validateObject,
-} = require('internal/validators');
-
-const {
   addAbortSignal,
 } = require('internal/streams/add-abort-signal');
-
-const {
-  queueMicrotask,
-} = require('internal/process/task_queues');
 
 const {
   getHighWaterMark,
@@ -76,20 +59,16 @@ const {
   AbortError,
   codes: {
     ERR_INVALID_ARG_TYPE,
-    ERR_INVALID_ARG_VALUE,
-    ERR_INVALID_STATE,
     ERR_METHOD_NOT_IMPLEMENTED,
     ERR_MULTIPLE_CALLBACK,
     ERR_STREAM_ALREADY_FINISHED,
     ERR_STREAM_CANNOT_PIPE,
     ERR_STREAM_DESTROYED,
-    ERR_STREAM_ITER_MISSING_FLAG,
     ERR_STREAM_NULL_VALUES,
     ERR_STREAM_WRITE_AFTER_END,
     ERR_UNKNOWN_ENCODING,
   },
 } = require('internal/errors');
-const { getOptionValue } = require('internal/options');
 const {
   kState,
   // bitfields
@@ -1179,438 +1158,3 @@ Writable.prototype[SymbolAsyncDispose] = async function() {
     eos(this, (err) => (err && err.name !== 'AbortError' ? reject(err) : resolve(null))),
   );
 };
-
-// Interop with the stream/iter API.
-//
-// toStreamIterWriter() returns a Writer-shaped adapter that bridges a classic
-// stream.Writable to the stream/iter Writer interface. Since all writes on
-// stream.Writable are fundamentally async, the sync methods (writeSync,
-// writevSync, endSync) return false / -1 to indicate they are not supported,
-// deferring to the async path.
-//
-// Like the Readable interop, the flag cannot be checked at module load time.
-// toStreamIterWriter is always defined but lazily initializes on first call.
-{
-  let drainableProtocol;
-  let validateBackpressure;
-  let toUint8Array;
-
-  // Only one Writer adapter per Writable instance. The adapter holds
-  // mutable state (totalBytes, waiters, listeners) that would conflict
-  // if two adapters existed for the same underlying stream.
-  const kStreamIterWriter = Symbol('kStreamIterWriter');
-
-  function lazyInit() {
-    if (drainableProtocol !== undefined) return;
-    if (!getOptionValue('--experimental-stream-iter')) {
-      throw new ERR_STREAM_ITER_MISSING_FLAG();
-    }
-    ({ drainableProtocol } = require('internal/streams/iter/types'));
-    ({
-      validateBackpressure,
-      toUint8Array,
-    } = require('internal/streams/iter/utils'));
-  }
-
-  Writable.prototype.toStreamIterWriter =
-      function toStreamIterWriter(options = { __proto__: null }) {
-        lazyInit();
-
-        if (this[kStreamIterWriter]) {
-          return this[kStreamIterWriter];
-        }
-
-        // The Writer interface is bytes-only. Object-mode Writables expect
-        // arbitrary JS values, which is incompatible.
-        if (this.writableObjectMode) {
-          throw new ERR_INVALID_STATE(
-            'Cannot create a stream/iter Writer from an object-mode Writable');
-        }
-
-        validateObject(options, 'options');
-        const {
-          backpressure = 'strict',
-        } = options;
-        validateBackpressure(backpressure);
-
-        // drop-oldest is not supported for classic stream.Writable. The
-        // Writable's internal buffer stores individual { chunk, encoding,
-        // callback } entries with no concept of batch boundaries. A writev()
-        // call fans out into N separate buffer entries, so a subsequent
-        // drop-oldest eviction could partially tear apart an earlier atomic
-        // writev batch. The PushWriter avoids this because writev occupies a
-        // single slot. Supporting drop-oldest here would require either
-        // accepting partial writev eviction or adding batch tracking to the
-        // buffer -- neither is acceptable without a deeper rework of Writable
-        // internals.
-        if (backpressure === 'drop-oldest') {
-          throw new ERR_INVALID_ARG_VALUE('options.backpressure', backpressure,
-                                          'drop-oldest is not supported for classic stream.Writable');
-        }
-
-        const writable = this;
-        const hwm = writable.writableHighWaterMark;
-        let totalBytes = 0;
-
-        // Waiters pending on backpressure resolution (block policy only).
-        // Multiple un-awaited writes can each add a waiter, so this must be
-        // a list. A single persistent 'drain' listener and 'error' listener
-        // (installed once lazily) resolve or reject all waiters to avoid
-        // accumulating per-write listeners on the stream.
-        let waiters = [];
-        let listenersInstalled = false;
-        let onDrain;
-        let onError;
-
-        function installListeners() {
-          if (listenersInstalled) return;
-          listenersInstalled = true;
-          onDrain = () => {
-            const pending = waiters;
-            waiters = [];
-            for (let i = 0; i < pending.length; i++) {
-              pending[i].resolve();
-            }
-          };
-          onError = (err) => {
-            const pending = waiters;
-            waiters = [];
-            for (let i = 0; i < pending.length; i++) {
-              pending[i].reject(err);
-            }
-          };
-          writable.on('drain', onDrain);
-          writable.on('error', onError);
-        }
-
-        // Reject all pending waiters and remove the drain/error listeners.
-        function cleanup(err) {
-          const pending = waiters;
-          waiters = [];
-          for (let i = 0; i < pending.length; i++) {
-            pending[i].reject(err ?? new AbortError());
-          }
-          if (!listenersInstalled) return;
-          listenersInstalled = false;
-          writable.removeListener('drain', onDrain);
-          writable.removeListener('error', onError);
-        }
-
-        function waitForDrain() {
-          const { promise, resolve, reject } = PromiseWithResolvers();
-          waiters.push({ __proto__: null, resolve, reject });
-          installListeners();
-          return promise;
-        }
-
-        function isWritable() {
-          return !writable.destroyed &&
-             !writable.writableFinished &&
-             !writable.writableEnded;
-        }
-
-        function isFull() {
-          return writable.writableLength >= hwm;
-        }
-
-        const writer = {
-          __proto__: null,
-
-          get desiredSize() {
-            if (!isWritable()) return null;
-            return MathMax(0, hwm - writable.writableLength);
-          },
-
-          writeSync(chunk) {
-            return false;
-          },
-
-          writevSync(chunks) {
-            return false;
-          },
-
-          // Backpressure semantics: write() resolves when the data is accepted
-          // into the Writable's internal buffer, NOT when _write() has flushed
-          // it to the underlying resource. This matches the Writer spec -- the
-          // PushWriter resolves on buffer acceptance too. Classic Writable flow
-          // control works the same way: write rapidly until write() returns
-          // false, then wait for 'drain'. The _write callback is involved in
-          // backpressure indirectly -- 'drain' fires after callbacks drain the
-          // buffer below highWaterMark. Per-write errors from _write surface
-          // as 'error' events caught by our generic error handler, rejecting
-          // the next pending operation rather than the already-resolved one.
-          //
-          // The options.signal parameter from the Writer interface is ignored.
-          // Classic stream.Writable has no per-write abort signal support;
-          // cancellation should be handled at the pipeline level instead.
-          write(chunk) {
-            if (!isWritable()) {
-              return PromiseReject(new ERR_STREAM_WRITE_AFTER_END());
-            }
-
-            let bytes;
-            try {
-              bytes = toUint8Array(chunk);
-            } catch (err) {
-              return PromiseReject(err);
-            }
-
-            if (backpressure === 'strict' && isFull()) {
-              return PromiseReject(new ERR_INVALID_STATE.RangeError(
-                'Backpressure violation: buffer is full. ' +
-            'Await each write() call to respect backpressure.'));
-            }
-
-            if (backpressure === 'drop-newest' && isFull()) {
-              // Silently discard. Still count bytes for consistency with
-              // PushWriter, which counts dropped bytes in totalBytes.
-              totalBytes += TypedArrayPrototypeGetByteLength(bytes);
-              return PromiseResolve();
-            }
-
-            totalBytes += TypedArrayPrototypeGetByteLength(bytes);
-            const ok = writable.write(bytes);
-            if (ok) return PromiseResolve();
-
-            // backpressure === 'block' (or strict with room that filled on
-            // this write -- writable.write() accepted the data but returned
-            // false indicating the buffer is now at/over hwm).
-            if (backpressure === 'block') {
-              return waitForDrain();
-            }
-
-            // strict: the write was accepted (there was room before writing)
-            // but the buffer is now full. Resolve -- the *next* write will
-            // be rejected if the caller ignores backpressure.
-            return PromiseResolve();
-          },
-
-          writev(chunks) {
-            if (!ArrayIsArray(chunks)) {
-              throw new ERR_INVALID_ARG_TYPE('chunks', 'Array', chunks);
-            }
-            if (!isWritable()) {
-              return PromiseReject(new ERR_STREAM_WRITE_AFTER_END());
-            }
-
-            if (backpressure === 'strict' && isFull()) {
-              return PromiseReject(new ERR_INVALID_STATE.RangeError(
-                'Backpressure violation: buffer is full. ' +
-            'Await each write() call to respect backpressure.'));
-            }
-
-            if (backpressure === 'drop-newest' && isFull()) {
-              // Discard entire batch.
-              for (let i = 0; i < chunks.length; i++) {
-                totalBytes +=
-              TypedArrayPrototypeGetByteLength(toUint8Array(chunks[i]));
-              }
-              return PromiseResolve();
-            }
-
-            writable.cork();
-            let ok = true;
-            for (let i = 0; i < chunks.length; i++) {
-              const bytes = toUint8Array(chunks[i]);
-              totalBytes += TypedArrayPrototypeGetByteLength(bytes);
-              ok = writable.write(bytes);
-            }
-            writable.uncork();
-
-            if (ok) return PromiseResolve();
-
-            if (backpressure === 'block') {
-              return waitForDrain();
-            }
-
-            return PromiseResolve();
-          },
-
-          endSync() {
-            return -1;
-          },
-
-          // options.signal is ignored for the same reason as write().
-          end() {
-            if (writable.writableFinished || writable.destroyed) {
-              cleanup();
-              return PromiseResolve(totalBytes);
-            }
-
-            const { promise, resolve, reject } = PromiseWithResolvers();
-
-            if (!writable.writableEnded) {
-              writable.end();
-            }
-
-            eos(writable, { writable: true, readable: false }, (err) => {
-              cleanup(err);
-              if (err) reject(err);
-              else resolve(totalBytes);
-            });
-
-            return promise;
-          },
-
-          fail(reason) {
-            cleanup(reason);
-            writable.destroy(reason);
-          },
-
-          [SymbolAsyncDispose]() {
-            if (isWritable()) {
-              cleanup();
-              writable.destroy();
-            }
-            return PromiseResolve();
-          },
-
-          [SymbolDispose]() {
-            if (isWritable()) {
-              cleanup();
-              writable.destroy();
-            }
-          },
-        };
-
-        // drainableProtocol is set after lazyInit
-        writer[drainableProtocol] = function() {
-          if (!isWritable()) return null;
-          if (writable.writableLength < hwm) {
-            return PromiseResolve(true);
-          }
-          const { promise, resolve } = PromiseWithResolvers();
-          waiters.push({
-            __proto__: null,
-            resolve() { resolve(true); },
-            reject() { resolve(false); },
-          });
-          installListeners();
-          return promise;
-        };
-
-        this[kStreamIterWriter] = writer;
-        return writer;
-      };
-
-  // Create a classic stream.Writable backed by a stream/iter Writer.
-  // Each _write/_writev call delegates to the Writer's async methods,
-  // attempting the sync path first (writeSync/writevSync/endSync) and
-  // falling back to async if the sync path returns false or throws.
-  // _final calls writer.end(), _destroy calls writer.fail().
-  Writable.fromStreamIter = function fromStreamIter(writer) {
-    lazyInit();
-
-    if (typeof writer?.write !== 'function') {
-      throw new ERR_INVALID_ARG_TYPE('writer', 'Writer', writer);
-    }
-
-    const hasWriteSync = typeof writer.writeSync === 'function';
-    const hasWritev = typeof writer.writev === 'function';
-    const hasWritevSync = hasWritev &&
-                          typeof writer.writevSync === 'function';
-    const hasEnd = typeof writer.end === 'function';
-    const hasEndSync = hasEnd &&
-                       typeof writer.endSync === 'function';
-    const hasFail = typeof writer.fail === 'function';
-
-    // Try-sync-first pattern: attempt the synchronous method and
-    // fall back to the async method if it returns false (indicating
-    // the sync path was not accepted) or throws. When the sync path
-    // succeeds, the callback is deferred via queueMicrotask to
-    // preserve the async resolution contract that Writable internals
-    // expect from _write/_writev/_final callbacks.
-
-    function _write(chunk, encoding, cb) {
-      const bytes = typeof chunk === 'string' ?
-        Buffer.from(chunk, encoding) : chunk;
-      if (hasWriteSync) {
-        try {
-          if (writer.writeSync(bytes)) {
-            queueMicrotask(cb);
-            return;
-          }
-        } catch {
-          // Sync path threw -- fall through to async.
-        }
-      }
-      try {
-        PromisePrototypeThen(writer.write(bytes), () => cb(), cb);
-      } catch (err) {
-        cb(err);
-      }
-    }
-
-    function _writev(entries, cb) {
-      const chunks = [];
-      for (let i = 0; i < entries.length; i++) {
-        const { chunk, encoding } = entries[i];
-        chunks[i] = typeof chunk === 'string' ?
-          Buffer.from(chunk, encoding) : chunk;
-      }
-      if (hasWritevSync) {
-        try {
-          if (writer.writevSync(chunks)) {
-            queueMicrotask(cb);
-            return;
-          }
-        } catch {
-          // Sync path threw -- fall through to async.
-        }
-      }
-      try {
-        PromisePrototypeThen(writer.writev(chunks), () => cb(), cb);
-      } catch (err) {
-        cb(err);
-      }
-    }
-
-    function _final(cb) {
-      if (!hasEnd) {
-        queueMicrotask(cb);
-        return;
-      }
-      if (hasEndSync) {
-        try {
-          const result = writer.endSync();
-          if (result >= 0) {
-            queueMicrotask(cb);
-            return;
-          }
-        } catch {
-          // Sync path threw -- fall through to async.
-        }
-      }
-      try {
-        PromisePrototypeThen(writer.end(), () => cb(), cb);
-      } catch (err) {
-        cb(err);
-      }
-    }
-
-    function _destroy(err, cb) {
-      if (err && hasFail) {
-        writer.fail(err);
-      }
-      cb();
-    }
-
-    const writableOptions = {
-      __proto__: null,
-      // Use MAX_SAFE_INTEGER to effectively disable the Writable's
-      // internal buffering. The underlying stream/iter Writer has its
-      // own backpressure handling; we want _write to be called
-      // immediately so the Writer can manage flow control directly.
-      highWaterMark: NumberMAX_SAFE_INTEGER,
-      write: _write,
-      final: _final,
-      destroy: _destroy,
-    };
-
-    if (hasWritev) {
-      writableOptions.writev = _writev;
-    }
-
-    return new Writable(writableOptions);
-  };
-}
