+++ lib/stream/iter.js
@@ -50,6 +50,15 @@ const {
   ondrain,
 } = require('internal/streams/iter/consumers');
 
+// Classic stream interop (Node.js-specific, not part of the spec)
+const {
+  fromReadable,
+  fromWritable,
+  toReadable,
+  toReadableSync,
+  toWritable,
+} = require('internal/streams/iter/classic');
+
 // Multi-consumer
 const { broadcast, Broadcast } = require('internal/streams/iter/broadcast');
 const {
@@ -177,4 +186,11 @@ module.exports = {
   tap,
   tapSync,
   ondrain,
+
+  // Classic stream interop
+  fromReadable,
+  fromWritable,
+  toReadable,
+  toReadableSync,
+  toWritable,
 };
