+++ test/parallel/test-stream-iter-readable-interop-disabled.js
@@ -38,35 +38,20 @@ async function testToAsyncStreamableWithFlag() {
   assert.strictEqual(code, 0);
 }
 
-async function testFromStreamIterWithoutFlag() {
-  const { stderr, code } = await spawnPromisified(process.execPath, [
+async function testStreamIterModuleWithoutFlag() {
+  // Requiring 'stream/iter' without the flag should not be possible
+  // since the module is gated behind --experimental-stream-iter.
+  const { code } = await spawnPromisified(process.execPath, [
     '-e',
     `
-      const { Readable } = require('stream');
-      async function* gen() { yield [Buffer.from('x')]; }
-      Readable.fromStreamIter(gen());
+      require('stream/iter');
     `,
   ]);
   assert.notStrictEqual(code, 0);
-  assert.match(stderr, /ERR_STREAM_ITER_MISSING_FLAG/);
-}
-
-async function testFromStreamIterSyncWithoutFlag() {
-  const { stderr, code } = await spawnPromisified(process.execPath, [
-    '-e',
-    `
-      const { Readable } = require('stream');
-      function* gen() { yield [Buffer.from('x')]; }
-      Readable.fromStreamIterSync(gen());
-    `,
-  ]);
-  assert.notStrictEqual(code, 0);
-  assert.match(stderr, /ERR_STREAM_ITER_MISSING_FLAG/);
 }
 
 Promise.all([
   testToAsyncStreamableWithoutFlag(),
   testToAsyncStreamableWithFlag(),
-  testFromStreamIterWithoutFlag(),
-  testFromStreamIterSyncWithoutFlag(),
+  testStreamIterModuleWithoutFlag(),
 ]).then(common.mustCall());
